Modify user edit page
