+++ app/controllers/user/info_controller.rb
@@ -1,5 +1,10 @@
 class User::InfoController < ApplicationController
+  before_action :authenticate_user!, except: :show
+
   def show
     @user = User.find_by!(username: params[:username])
+  end
+
+  def edit
   end
 end
