+++ app/javascript/packs/gif_form.js
@@ -22,16 +22,3 @@ remote_url_field.addEventListener('input', e => {
 preview.addEventListener('click', () => {
   file_field.click();
 });
-
-// Tagify
-import Tagify from '@yaireo/tagify'
-new Tagify(document.querySelector('[name="gif[tag_list]"'));
-
-// Submit
-const form = document.querySelector('form');
-const submit_button = document.querySelector('#submit_button');
-submit_button.addEventListener('click', e => {
-  submit_button.setAttribute('disabled', '');
-  submit_button.textContent = 'Wait...';
-  form.submit();
-});
