+++ app/views/gifs/show.html.erb
@@ -3,16 +3,16 @@
 
 <%= image_tag @gif.image %>
 
-<p class="title is-3">
-  <%= title @gif.title %>
+<div class="content">
+  <p class="title is-3 is-inline"><%= title @gif.title %></p>
   <% if @gif.editable?(current_user) %>
-    <%= link_to edit_gif_path(@gif) do %>
+    <%= link_to edit_gif_path(@gif), class: 'is-inline' do %>
       <span class="icon">
         <i class="fas fa-edit"></i>
       </span>
     <% end %>
   <% end %>
-</p>
+</div>
 
 <div class="field has-addons copy-button">
   <div class="control has-icons-right">
@@ -36,7 +36,7 @@
   </a>
 </div>
 
-<div class="padding-20 content">
+<div class="content">
   <p>
     <span class="icon">
       <i class="fas fa-share-square"></i>
